+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/SessionV2/OnlineAsyncTaskAccelByteInitializePlayerAttributes.h
@@ -19,6 +19,7 @@ public:
 
 	virtual void Initialize() override;
 	virtual void Finalize() override;
+	virtual void TriggerDelegates() override;
 
 protected: