+++ Source/OnlineSubsystemAccelByte/Public/OnlineSubsystemAccelByte.h
@@ -12,7 +12,6 @@
 #include "OnlineEntitlementsInterfaceAccelByte.h"
 #include "OnlinePurchaseInterfaceAccelByte.h"
 #include "OnlineStoreInterfaceV2AccelByte.h"
-#include "OnlineAnalyticsInterfaceAccelByte.h"
 #include "OnlineChatInterfaceAccelByte.h"
 #include "Core/AccelByteApiClient.h"
 #include "Models/AccelByteUserModels.h"